Restructed app.py, created routes.py
